Run doctests for __main__.py entrypoints
they are ignored by pytest in recent versions.
